Remove old drivers
Updated version of #592.

1. Move DYCOMSRadiation and DYCOMS_BC into driver
2. Migrate Rayleigh-Benard test + b.c. from test/DGmethods into driver-config
   and into examples/Atmos with default-override comments
